fix long to int

+++ entropy_data.cs
@@ -11,7 +11,7 @@ using System;
 
 class entropy_data
 {
-	internal long slow_level;
+	internal int slow_level;
 	internal int[] median = new int[]{0, 0, 0}; // was uint32_t in C, we initialize in order to remove run time errors
-	internal long error_limit; // was uint32_t in C
+	internal int error_limit; // was uint32_t in C
 }
